backend/requirements.txt
fastapi
uvicorn[standard]
easyocr
pydantic
httpx
pillow
